Moved demos resources to engine/desktop
