+++ engine/desktop/src/test/java/es/eucm/ead/engine/demos/SpineDemo.java
@@ -36,14 +36,14 @@
  */
 package es.eucm.ead.engine.demos;
 
-import es.eucm.ead.editor.demobuilder.EditorDemoBuilder;
+import es.eucm.ead.engine.demobuilder.ExecutableDemoBuilder;
 import es.eucm.ead.schema.entities.ModelEntity;
 import es.eucm.ead.schema.renderers.SpineAnimation;
 
 /**
  * Created by angel on 29/11/14.
  */
-public class SpineDemo extends EditorDemoBuilder {
+public class SpineDemo extends ExecutableDemoBuilder {
 
 	public SpineDemo() {
 		super("spine");
