+++ engine/desktop/src/test/resources/planes-demo.png
