+++ engine/desktop/src/test/resources/spine.png
