+++ engine/desktop/src/test/resources/spine.zip
